requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
async-timeout==4.0.3
    # via redis
blinker==1.7.0
    # via flask
certifi==2024.2.2
    # via
    #   polygon-api-client
    #   requests
celery
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
flask==3.0.3
    # via -r requirements.in
idna==3.7
    # via requests
itsdangerous==2.2.0
holidays
    # via flask
jinja2==3.1.3
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
polygon-api-client==1.13.6
    # via -r requirements.in
python-dotenv
redis==5.0.4
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
urllib3==1.26.18
    # via
    #   polygon-api-client
    #   requests
websockets==12.0
    # via polygon-api-client
werkzeug==3.0.2
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
